Various sample drawing fixups
Mostly fixing stuff I'd broken earlier in the text work, and hadn't
noticed.
