+++ piet/src/samples/picture_5.rs
@@ -1,61 +1,41 @@
-//! Basic example of just text
+//! range attributes should override default attributes
 
-use crate::kurbo::{Line, Rect, Size};
+use crate::kurbo::{Size, Vec2};
 use crate::{
-    Color, Error, FontBuilder, RenderContext, Text, TextAttribute, TextLayout, TextLayoutBuilder,
+    Color, Error, FontBuilder, FontWeight, RenderContext, Text, TextAttribute, TextLayout,
+    TextLayoutBuilder,
 };
 
-pub const SIZE: Size = Size::new(400., 200.);
-pub const TEXT_COLOR: Color = Color::rgba8(0x00, 0x80, 0x80, 0xF0);
+pub const SIZE: Size = Size::new(480., 560.);
 
-pub fn draw(rc: &mut impl RenderContext) -> Result<(), Error> {
-    // Black background
-    rc.clear(Color::BLACK);
+static TEXT: &str = r#"Philosophers often behave like little children who scribble some marks on a piece of paper at random and then ask the grown-up "What's that?" — It happened like this: the grown-up had drawn pictures for the child several times and said "this is a man," "this is a house," etc. And then the child makes some marks too and asks: what's this then?"#;
 
-    // do something texty
-    let font = rc.text().new_font_by_name("Segoe UI", 12.0).build()?;
+const RED: Color = Color::rgb8(255, 0, 0);
+const BLUE: Color = Color::rgb8(0, 0, 255);
 
-    let layout = rc
-        .text()
-        .new_text_layout(&font, "piet text!", None)
-        .default_attribute(TextAttribute::ForegroundColor(TEXT_COLOR))
+pub fn draw<R: RenderContext>(rc: &mut R) -> Result<(), Error> {
+    rc.clear(Color::WHITE);
+    let text = rc.text();
+    let font = text.system_font(12.0);
+    let font2 = text.new_font_by_name("Courier New", 12.0).build().unwrap();
+    let layout = text
+        .new_text_layout(&font, TEXT, 200.0)
+        .default_attribute(font2)
+        .default_attribute(TextAttribute::Underline(true))
+        .default_attribute(TextAttribute::Italic(true))
+        .default_attribute(TextAttribute::ForegroundColor(RED))
+        .default_attribute(FontWeight::BOLD)
+        .range_attribute(..200, TextAttribute::ForegroundColor(BLUE))
+        .range_attribute(10..100, FontWeight::NORMAL)
+        .range_attribute(40..300, TextAttribute::Underline(false))
+        .range_attribute(60..160, TextAttribute::Italic(false))
+        .range_attribute(140..220, FontWeight::NORMAL)
+        .range_attribute(240.., font)
         .build()?;
 
-    let layout_multiline = rc
-        .text()
-        .new_text_layout(&font, "piet text is the best text!", 50.0)
-        .default_attribute(TextAttribute::ForegroundColor(TEXT_COLOR))
-        .build()?;
-
-    let width = layout.size().width;
-
-    let multiline_bg = Rect::from_origin_size((20.0, 50.0), (50.0, 100.0));
-    rc.fill(multiline_bg, &Color::rgb(0.3, 0.0, 0.4));
-
-    rc.draw_text(&layout, (100.0, 50.0));
-    rc.draw_text(&layout_multiline, (20.0, 50.0));
-
-    // underline text
-    rc.stroke(
-        Line::new((100.0, 52.0), (100.0 + width, 52.0)),
-        &TEXT_COLOR,
-        1.0,
-    );
-
-    // get hit test text position and draw a cursor
-    let hit_test_text_position = layout.hit_test_text_position(3);
-
-    if let Some(http) = hit_test_text_position {
-        let cursor_x = http.point.x;
-
-        let color = Color::rgba8(0x80, 0x80, 0x80, 0xF0);
-
-        rc.stroke(
-            Line::new((100.0 + cursor_x, 50.0), (100.0 + cursor_x, 50.0 - 10.0)),
-            &color,
-            1.0,
-        );
-    }
+    let y_pos = ((SIZE.height - layout.size().height * 2.0) / 4.0).max(0.0);
+    let text_pos = Vec2::new(16.0, y_pos);
+    rc.draw_text(&layout, text_pos.to_point());
 
     Ok(())
 }
